clean up, refactor into different mod/files
