+++ dkg-gadget/src/async_protocols/mod.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 pub mod blockchain_interface;
-pub mod dkg;
 pub mod incoming;
 pub mod keygen;
 pub mod remote;
