+++ dkg-gadget/src/dkg_modules/mp_ecdsa.rs
@@ -1,84 +1,21 @@
 use crate::{
-	async_protocols::{
-		remote::AsyncProtocolRemote, GenericAsyncHandler, KeygenPartyId, KeygenRound,
+	async_protocols::{GenericAsyncHandler, KeygenRound},
+	dkg_modules::{
+		KeygenProtocolSetupParameters, ProtocolInitReturn, SigningProtocolSetupParameters, DKG,
 	},
 	gossip_engine::GossipEngineIface,
 	keygen_manager::KeygenState,
-	utils::SendFuture,
 	worker::{DKGWorker, ProtoStageType},
 	Client,
 };
 use async_trait::async_trait;
 use dkg_primitives::types::DKGError;
-use dkg_runtime_primitives::{
-	crypto::{AuthorityId, Public},
-	BatchId, DKGApi, MaxAuthorities, MaxProposalLength, MaxProposalsInBatch, SessionId,
-	StoredUnsignedProposalBatch,
-};
-use parking_lot::RwLock;
+use dkg_runtime_primitives::{crypto::AuthorityId, DKGApi, MaxAuthorities, MaxProposalLength};
 use sc_client_api::Backend;
 use sp_runtime::traits::{Block, NumberFor};
-use std::{
-	marker::PhantomData,
-	pin::Pin,
-	sync::{atomic::Ordering, Arc},
-};
+use std::sync::atomic::Ordering;
 
-pub enum DKGProtocolSetupParameters<B: Block> {
-	MpEcdsa {
-		best_authorities: Vec<(KeygenPartyId, Public)>,
-		authority_public_key: Public,
-		party_i: KeygenPartyId,
-		session_id: SessionId,
-		associated_block: NumberFor<B>,
-		threshold: u16,
-		stage: ProtoStageType,
-		keygen_protocol_hash: [u8; 32],
-	},
-	WTFrost {},
-}
-
-pub enum SigningProtocolSetupParameters<B: Block> {
-	MpEcdsa {
-		best_authorities: Vec<(KeygenPartyId, Public)>,
-		authority_public_key: Public,
-		party_i: KeygenPartyId,
-		session_id: SessionId,
-		threshold: u16,
-		stage: ProtoStageType,
-		unsigned_proposal_batch: StoredUnsignedProposalBatch<
-			BatchId,
-			MaxProposalLength,
-			MaxProposalsInBatch,
-			NumberFor<B>,
-		>,
-		signing_set: Vec<KeygenPartyId>,
-		associated_block_id: NumberFor<B>,
-	},
-	WTFrost {},
-}
-
-/// A type which is used directly by the Job Manager to initialize and manage the DKG protocol
-pub type ProtocolInitReturn<B> =
-	(AsyncProtocolRemote<NumberFor<B>>, Pin<Box<dyn SendFuture<'static, ()>>>);
-
-#[async_trait]
-/// Generalizes the DKGWorker::initialize_keygen_protocol and DKGWorker::initialize_signing_protocol
-/// Also includes two functions which are used for determining whether a DKG can handle the request
-/// parameters, which is used in the [`DKGModules`] implementation
-pub trait DKG<B: Block>: Send + Sync {
-	async fn initialize_keygen_protocol(
-		&self,
-		params: DKGProtocolSetupParameters<B>,
-	) -> Option<ProtocolInitReturn<B>>;
-	async fn initialize_signing_protocol(
-		&self,
-		params: SigningProtocolSetupParameters<B>,
-	) -> Result<ProtocolInitReturn<B>, DKGError>;
-	fn can_handle_keygen_request(&self, params: &DKGProtocolSetupParameters<B>) -> bool;
-	fn can_handle_signing_request(&self, params: &SigningProtocolSetupParameters<B>) -> bool;
-}
-
+/// DKG module for Multi-Party ECDSA
 pub struct MpEcdsaDKG<B, BE, C, GE>
 where
 	B: Block,
@@ -86,7 +23,7 @@ where
 	C: Client<B, BE>,
 	GE: GossipEngineIface,
 {
-	dkg_worker: DKGWorker<B, BE, C, GE>,
+	pub(super) dkg_worker: DKGWorker<B, BE, C, GE>,
 }
 
 #[async_trait]
@@ -100,9 +37,9 @@ where
 {
 	async fn initialize_keygen_protocol(
 		&self,
-		params: DKGProtocolSetupParameters<B>,
+		params: KeygenProtocolSetupParameters<B>,
 	) -> Option<ProtocolInitReturn<B>> {
-		if let DKGProtocolSetupParameters::MpEcdsa {
+		if let KeygenProtocolSetupParameters::MpEcdsa {
 			best_authorities,
 			authority_public_key,
 			party_i,
@@ -259,104 +196,11 @@ where
 		}
 	}
 
-	fn can_handle_keygen_request(&self, params: &DKGProtocolSetupParameters<B>) -> bool {
-		matches!(params, DKGProtocolSetupParameters::MpEcdsa { .. })
+	fn can_handle_keygen_request(&self, params: &KeygenProtocolSetupParameters<B>) -> bool {
+		matches!(params, KeygenProtocolSetupParameters::MpEcdsa { .. })
 	}
 
 	fn can_handle_signing_request(&self, params: &SigningProtocolSetupParameters<B>) -> bool {
 		matches!(params, SigningProtocolSetupParameters::MpEcdsa { .. })
-	}
-}
-
-pub struct WTFrostDKG {}
-
-#[async_trait]
-impl<B: Block> DKG<B> for WTFrostDKG {
-	async fn initialize_keygen_protocol(
-		&self,
-		_params: DKGProtocolSetupParameters<B>,
-	) -> Option<ProtocolInitReturn<B>> {
-		todo!()
-	}
-
-	async fn initialize_signing_protocol(
-		&self,
-		_params: SigningProtocolSetupParameters<B>,
-	) -> Result<ProtocolInitReturn<B>, DKGError> {
-		todo!()
-	}
-
-	fn can_handle_keygen_request(&self, params: &DKGProtocolSetupParameters<B>) -> bool {
-		matches!(params, DKGProtocolSetupParameters::WTFrost { .. })
-	}
-
-	fn can_handle_signing_request(&self, params: &SigningProtocolSetupParameters<B>) -> bool {
-		matches!(params, SigningProtocolSetupParameters::WTFrost { .. })
-	}
-}
-
-/// Holds a list of DKGs that can be used at runtime
-pub struct DKGModules<B: Block, BE, C, GE> {
-	dkgs: Arc<RwLock<Vec<Arc<dyn DKG<B>>>>>,
-	_pd: PhantomData<(BE, C, GE)>,
-}
-
-impl<B, BE, C, GE> DKGModules<B, BE, C, GE>
-where
-	B: Block,
-	BE: Backend<B> + Unpin + 'static,
-	C: Client<B, BE> + 'static,
-	GE: GossipEngineIface,
-	C::Api: DKGApi<B, AuthorityId, NumberFor<B>, MaxProposalLength, MaxAuthorities>,
-{
-	pub fn initialize(&self, dkg_worker: DKGWorker<B, BE, C, GE>) {
-		*self.dkgs.write() = vec![Arc::new(MpEcdsaDKG { dkg_worker }), Arc::new(WTFrostDKG {})]
-	}
-
-	pub fn get_keygen_protocol(
-		&self,
-		params: &DKGProtocolSetupParameters<B>,
-	) -> Option<Arc<dyn DKG<B>>> {
-		self.dkgs
-			.read()
-			.iter()
-			.find(|dkg| dkg.can_handle_keygen_request(params))
-			.cloned()
-	}
-
-	pub fn get_signing_protocol(
-		&self,
-		params: &SigningProtocolSetupParameters<B>,
-	) -> Option<Arc<dyn DKG<B>>> {
-		self.dkgs
-			.read()
-			.iter()
-			.find(|dkg| dkg.can_handle_signing_request(params))
-			.cloned()
-	}
-}
-
-// implement Default for DKGModules
-impl<B, BE, C, GE> Default for DKGModules<B, BE, C, GE>
-where
-	B: Block,
-	BE: Backend<B> + 'static,
-	C: Client<B, BE> + 'static,
-	GE: GossipEngineIface,
-{
-	fn default() -> Self {
-		Self { dkgs: Arc::new(RwLock::new(vec![])), _pd: PhantomData }
-	}
-}
-
-impl<B, BE, C, GE> Clone for DKGModules<B, BE, C, GE>
-where
-	B: Block,
-	BE: Backend<B> + 'static,
-	C: Client<B, BE> + 'static,
-	GE: GossipEngineIface,
-{
-	fn clone(&self) -> Self {
-		Self { dkgs: self.dkgs.clone(), _pd: PhantomData }
 	}
 }
